requirements.txt
absl-py==1.4.0
attrs==22.2.0
autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5d968ni_yn/croot/certifi_1671487774636/work/certifi
contourpy==1.0.7
cycler==0.11.0
flatbuffers==23.3.3
fonttools==4.39.3
importlib-resources==5.12.0
install==1.3.5
kiwisolver==1.4.4
matplotlib==3.7.1
mediapipe-silicon==0.9.1
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.0
paho-mqtt==1.6.1
Pillow==9.4.0
protobuf==3.20.3
pycodestyle @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5b2mq44vl0/croot/pycodestyle_1674267228581/work
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
zipp==3.15.0
